first warning fixes 1-17

+++ Userland/PinkOS/audioLib.c
@@ -6,7 +6,7 @@
 extern void syscall(uint64_t syscall, uint64_t arg1, uint64_t arg2, uint64_t arg3, uint64_t arg4, uint64_t arg5);
 
 void play_audio(Note** notes, uint8_t loop, uint64_t tempo){
-    syscall(PLAY_AUDIO_SYSCALL, (uint64_t)notes, loop, tempo, 0, 0);
+    syscall(PLAY_AUDIO_SYSCALL, (uint64_t)notes, (uint64_t)loop, tempo, 0, 0);
 }
 
 void stop_audio(){
@@ -33,8 +33,6 @@ void load_audio_state(AudioState state){
 
 int is_audio_playing(){
     int is_playing;
-    syscall(IS_AUDIO_PLAYING_SYSCALL, &is_playing, 0, 0, 0, 0);
+    syscall(IS_AUDIO_PLAYING_SYSCALL, (uint64_t)&is_playing, 0, 0, 0, 0);
     return is_playing;
 }
-
-
