+++ Userland/PinkOS/programs/date.c
@@ -15,7 +15,7 @@ extern void syscall(uint64_t syscall, uint64_t arg1, uint64_t arg2, uint64_t arg
 
 void date_main(unsigned char * args){
     RTC_Time time;
-    syscall(GET_RTC_TIME_SYSCALL, &time, 0, 0, 0, 0);
+    syscall(GET_RTC_TIME_SYSCALL, (uint64_t)&time, 0, 0, 0, 0);
     unsigned char * month;
     unsigned char * weekday;
     switch (time.month){
